requirements.txt
python-telegram-bot==21.0.1
yookassa==3.0.0
python-dotenv==1.0.0
flask==2.3.3
